IA_Alumno/requirements.txt
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
python-dotenv>=1.0.0
requests>=2.25.1